requirements.txt
flet
lxml
requests
pillow
pywin32; sys_platform == 'win32'
